pyproject.toml
[tool.poetry]
name = "mbcet-workshop"
version = "0.1.0"
description = ""
authors = ["Kesia Mary Joies <kesia@strollby.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
python-decouple = "^3.8"
mongoengine = "^0.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
